refactor: decompose db/schema.ts into DDL, migrations, and schema modules
schema.ts was 1,269 lines with three concerns mixed together: DDL constants,
8 migration functions, and the createSchema orchestrator with repetitive
version-check boilerplate. Split into three focused modules:

- schema-ddl.ts (653 lines): all table/FTS/index DDL constants
- migrations.ts (536 lines): migration functions + declarative MIGRATIONS registry
- schema.ts (83 lines): SCHEMA_VERSION, EMBEDDING_DIMENSIONS, createSchema loop

The migration chain now uses a simple for-of loop over the MIGRATIONS registry
instead of 8 hand-coded version-check blocks.
